# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Issue Tracker API
  description: REST API for a Redmine-like issue tracker. Supports projects, issues, comments, and role-based access control.
  license:
    name: ''
  version: 0.1.0
paths:
  /api/auth/login:
    post:
      tags:
      - Auth
      summary: Log in
      description: Authenticates with email and password, returns an access/refresh token pair.
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/refresh:
    post:
      tags:
      - Auth
      summary: Refresh tokens
      description: Validates a refresh token and issues a new access/refresh token pair.
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/register:
    post:
      tags:
      - Auth
      summary: Register a new user
      description: Creates a new user account and returns an access/refresh token pair.
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/comments/{comment_id}:
    delete:
      tags:
      - Comments
      summary: Delete comment
      description: Deletes a comment. Only the comment author can perform this action.
      operationId: delete_comment
      parameters:
      - name: comment_id
        in: path
        description: Comment ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Comment deleted
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied (not author)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer: []
...
  /api/issues/{issue_id}/comments:
    get:
      tags:
      - Comments
      summary: List comments
      description: Lists comments for an issue. Requires visibility access to the parent project.
      operationId: list_comments
      parameters:
      - name: issue_id
        in: path
        description: Issue ID
        required: true
        schema:
          type: string
          format: uuid
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Paginated list of comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCommentResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Issue not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer: []
    post:
      tags:
      - Comments
      summary: Create comment
      description: Creates a new comment on an issue. Requires membership in the parent project.
      operationId: create_comment
      parameters:
      - name: issue_id
        in: path
        description: Issue ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
        required: true
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Issue not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer: []
...
components:
  schemas:
...
    Comment:
      type: object
      required:
      - id
      - issue_id
      - author_id
      - body
      - created_at
      - updated_at
      properties:
        author_id:
          type: string
          format: uuid
        body:
          type: string
        created_at:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        issue_id:
          type: string
          format: uuid
        updated_at:
          type: string
          format: date-time
    CreateCommentRequest:
      type: object
      description: Request body for `POST /api/issues/:issue_id/comments`.
      required:
      - body
      properties:
        body:
          type: string
...
    ErrorResponse:
      type: object
      description: |-
        Schema-only type for documenting the standard error response shape in OpenAPI.
        Not used at runtime — the actual JSON is built inline in `AppError::into_response`.
      required:
      - message
      properties:
        data:
          description: Additional error details (validation field errors, etc.) or null.
        message:
          type: string
          description: Human-readable error message.
...
    LoginRequest:
      type: object
      description: Request body for `POST /api/auth/login`.
      required:
      - email
      - password
      properties:
        email:
          type: string
        password:
          type: string
...
    PaginatedCommentResponse:
      type: object
      required:
      - data
      - total
      - page
      - limit
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        limit:
          type: integer
          format: int64
        page:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
...
    RefreshRequest:
      type: object
      description: Request body for `POST /api/auth/refresh`.
      required:
      - refresh_token
      properties:
        refresh_token:
          type: string
    RegisterRequest:
      type: object
      description: Request body for `POST /api/auth/register`.
      required:
      - email
      - password
      - display_name
      properties:
        display_name:
          type: string
        email:
          type: string
        password:
          type: string
    TokenResponse:
      type: object
      description: Response returned after login or refresh.
      required:
      - access_token
      - refresh_token
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...